+++ src/app/redux/reducers.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import { viewConfigSlice } from './slices';
-
-export default combineReducers({
-  viewConfig: viewConfigSlice.reducer,
-});